Help bash know the prompt length
This fixes line editing glitches when composing a command at a
prompt that has the pipestatus display.

This may not be the best way to fix it, as the code is complicated
due to all the escaping to use `sed`, and as it prints invisible
escape characters even when they are not needed because the text
they would color is zero-length.